project: heartgold_soulsilver: fix CARDi_ReadCard custom overlays
The previous hijack only hijacked one of the two possible return statements in the CARDi_ReadCard function.
You could just also add a second hook in the other return statement, however, CARDi_ReadCard is actually
never called directly. Instead, there is a function that returns the address of this function, which
is then used to call it. We can edit the immediate containing the address of CARDi_ReadCard to our custom
payload instead, and then make sure to call CARDi_ReadCard ourselves in our payload.

This makes things a little cleaner and lets us drop the compiling of the hook seperately.

Signed-off-by: DevreeseJorik <jorik.devreese@outlook.com>